Make visICA, interp, visRej STEP2. bsfilt for line noise
In this, the trials visualized by ft_rejectvisual will be ONLY the bad
trials (although in the title it will show 'GOOD'). You have to go
through these and mark those you want to FULLY REMOVE (make nan), 'BAD' using the UI.
In the code this variable is 'veryBadTr'. Only the bad electrodes from the
leftover bad trials will be removed (made nan).